tests/requirements.txt
pytest
pytest-cov
pytest-mock
pytest-rerunfailures
pytest-xdist
pytest-snapshot>=0.8.1
Pillow==8.1.2
imgcat>=0.5.0
psutil==5.8.0
testpath
